app/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
qdrant-client>=1.7
numpy>=1.24
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.2
tensorflow>=2.15
